+++ R/ggadd_corr.R
@@ -1,6 +1,14 @@
 ggadd_corr <- function(p, resmca, var, cat=levels(var)[1], axes=c(1,2),
                        xbins=20, ybins=20, min.n=1, pal="RdYlBu", limits=NULL, legend="right") {
 
+  type <- attr(resmca,'class')[1]
+  
+  if(type=="stMCA") type <- resmca$call$input.mca
+  if(type=="csMCA") var <- var[resmca$call$subcloud]
+  if(type=="multiMCA") {
+    if(class(resmca$my.mca[[1]])[1]=="csMCA") var <- var[resmca$my.mca[[1]]$call$subcloud]
+  }
+  
   df <- as.data.frame(resmca$ind$coord[,axes])
   names(df) <- c("dim.1","dim.2")
   if(is.numeric(var)) df$vsup <- var