+++ NEWS.md
@@ -37,6 +37,7 @@
 * `conc.ellipse()` : several kinds of inertia ellipses can be plotted thanks to the kappa option
 * `ggadd_ellipses()` : level is set to 0.05 by default, which corresponds to conventional confidence ellipses. Option 'points' to choose to color the points or not.
 * `modif.rate()` : computes raw and modified rates
+* `homog.test()` : new dim argument
 * `modif.rate()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
 * `ggcloud_variables()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
 * `ggcloud_indiv()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
@@ -46,6 +47,12 @@
 * `dimcontrib()` : compatibility with objects of class MCA, speMCA and csMCA
 * `tabcontrib()` : compatibility with objects of class MCA, speMCA and csMCA
 * `homog.test()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
+* `varsup()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
+* `ggadd_chulls()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
+* `ggadd_corr()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
+* `ggadd_density()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
+* `ggadd_ellipses()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
+* `ggadd_kellipses()` : compatibility with objects of class MCA, speMCA, csMCA, stMCA and multiMCA
 
 ## Bug fixes